+++ src/middlewares/credential/index.ts
@@ -76,15 +76,3 @@ export const checkRefreshCredential = async (ctx: Context, next: Next) => {
 export const checkCredential = passport.authenticate('local', {
   session: false,
 });
-
-export const googleOAuth = passport.authenticate('google', {
-  session: false,
-  scope: ['profile', 'email'],
-  accessType: 'offline',
-  prompt: 'consent',
-});
-
-export const googleOAuthCallback = passport.authenticate('google', {
-  session: false,
-  failureRedirect: '/callback/failure',
-});
